Fix MCTS value estimation and backpropagation
Instead of the actual value estimate, the sum of value estimates was
being backpropagated. This works most of the time because the initial
sum is equal to the estimate and most nodes will only backpropagate
their value once right after expansion. However for final states,
this value can be backpropagated many times, causing sums to be added
 to sums. This grew without bound, eventually overflowing.

 In addition, the estimates were done on a child of the expanded
 leaf node. In combination with the initial value for E being set
 when expanding the leaf (which by itself would have caused it to be
 double the intended value after backpropagation) and then being used
 as the estimate, this caused the estimate to be 0 when not doing
 rollouts, since the child wasn't expanded.